+++ activerecord/test/cases/nested_attributes_test.rb
@@ -1064,4 +1064,27 @@ class TestHasManyAutosaveAssociationWhichItselfHasAutosaveAssociations < ActiveR
     assert_not part.valid?
     assert_equal ["Ship name can't be blank"], part.errors.full_messages
   end
+
+  class ProtectedParameters
+    def initialize(hash)
+      @hash = hash
+    end
+
+    def permitted?
+      true
+    end
+
+    def to_h
+      @hash
+    end
+  end
+
+  test "strong params style objects can be assigned" do
+    params = { name: "Stern", ship_attributes:
+               ProtectedParameters.new(name: "The Black Rock") }
+    part = ShipPart.new(params)
+
+    assert_equal "Stern", part.name
+    assert_equal "The Black Rock", part.ship.name
+  end
 end